code refactoring: moving some functions to utilities module
