+++ src/rules/rule/mod.rs
@@ -4,6 +4,8 @@ use network_object::NetworkObject;
 mod port_object;
 use port_object::PortObject;
 
+mod utilities;
+
 #[derive(Debug)]
 pub struct Rule {
     name: String,
