+++ src/rules/rule/network_object/mod.rs
@@ -6,6 +6,8 @@ use std::vec;
 use group::prefix_list::PrefixList;
 use group::Group;
 
+use super::utilities;
+
 #[derive(Debug)]
 pub struct NetworkObject {
     name: String,
@@ -30,6 +32,8 @@ pub enum NetworkObjectError {
     PrefixListError(#[from] group::prefix_list::PrefixListError),
     #[error("PANIC in parse network object: {0}")]
     Panic(String),
+    #[error("Fail to parse network object: {0}")]
+    NameExtractionError(#[from] utilities::UtilitiesError),
 }
 
 impl TryFrom<&Vec<String>> for NetworkObject {
@@ -51,7 +55,7 @@ impl TryFrom<&Vec<String>> for NetworkObject {
             ));
         }
 
-        let (name, merged_lines) = extract_name(lines)?;
+        let (name, merged_lines) = utilities::extract_name(lines)?;
 
         let mut items = vec![];
         let mut idx = 0;
@@ -134,89 +138,9 @@ fn get_lines_in_group(
     Ok(idx)
 }
 
-fn extract_name(lines: &[String]) -> Result<(String, Vec<String>), NetworkObjectError> {
-    if lines.is_empty() {
-        return Err(NetworkObjectError::General(
-            "Input lines are empty".to_string(),
-        ));
-    }
-
-    let first_line: Vec<_> = lines[0].split(": ").collect();
-    if first_line.len() != 2 {
-        return Err(NetworkObjectError::General2(
-            lines[0].to_string(),
-            "Incorrect first line format, expected '____ _____ : group or prefix'".to_string(),
-        ));
-    }
-    let name = first_line
-        .first()
-        .ok_or_else(|| {
-            NetworkObjectError::General2(
-                lines[0].to_string(),
-                "Missing name in first line".to_string(),
-            )
-        })?
-        .trim()
-        .to_string();
-    let merged_lines: Vec<_> = first_line[1..]
-        .iter()
-        .map(|x| x.to_string())
-        .chain(lines[1..].iter().map(|x| x.to_string()))
-        .collect();
-
-    Ok((name, merged_lines))
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn test_extract_name_valid() {
-        let lines = vec![
-            "Source Networks       : Internal (group)".to_string(),
-            "OBJ-10.11.12.0_23 (10.11.12.0/23)".to_string(),
-            "10.0.0.0/8".to_string(),
-        ];
-        let (name, merged_lines) = extract_name(&lines).unwrap();
-        assert_eq!(name, "Source Networks");
-        assert_eq!(
-            merged_lines,
-            vec![
-                "Internal (group)".to_string(),
-                "OBJ-10.11.12.0_23 (10.11.12.0/23)".to_string(),
-                "10.0.0.0/8".to_string(),
-            ]
-        );
-    }
-
-    #[test]
-    fn test_extract_name_invalid_format() {
-        let lines = vec!["Source Networks Internal (group)".to_string()];
-        let result = extract_name(&lines);
-        assert!(result.is_err());
-        if let Err(NetworkObjectError::General2(line, msg)) = result {
-            assert_eq!(line, "Source Networks Internal (group)");
-            assert_eq!(
-                msg,
-                "Incorrect first line format, expected '____ _____ : group or prefix'"
-            );
-        } else {
-            panic!("Expected NetworkObjectError::General2");
-        }
-    }
-
-    #[test]
-    fn test_extract_name_empty_lines() {
-        let lines: Vec<String> = vec![];
-        let result = extract_name(&lines);
-        assert!(result.is_err());
-        if let Err(NetworkObjectError::General(msg)) = result {
-            assert_eq!(msg, "Input lines are empty");
-        } else {
-            panic!("Expected NetworkObjectError::General");
-        }
-    }
 
     #[test]
     fn test_get_lines_in_group_single_group() {
